fix: some critical issues
